refactor(updates): split what Sparkle's UI must do from the seam
`AppUpdater.swift` reached 289 lines carrying two subjects: why
the update seam is a protocol with an inert default, and what
Sparkle's UI has to do differently for an app with no Dock tile.
§2.1 asks for the split before the ceiling rather than after, and
the second subject is the one that grew.

`UpdatePromptDriver.swift` takes the driver and its policy — what
they DECLARE. `AppUpdater.swift` keeps the seam, its two
conformers and the factory — where they are WIRED. The guard
follows the same cut: a needle asking what a class says reads the
first file, one asking what a call was handed reads the second.
